Refactor simulator into modular package
